moved app-bar from /src/components to /src/modules
